requirements.txt
playwright==1.47.0
pytest-playwright==0.5.2
python-dotenv==1.0.1